requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
scipy>=1.7.0
statsmodels>=0.13.0
lightgbm>=3.3.0
tensorflow>=2.10.0
torch>=1.12.0
tsaug>=0.1.0
pyyaml>=6.0
joblib>=1.1.0